Replace steebchen/nginx-spa:latest with own config
Steeb's Nginx configuration makes an assumption that the SPA will not
manage user routes containing dots (`.`). This is problematic as the dot
charater is part of the EOSIO name space (e.g. the `eosio.token`
account).

Thus, this commit replaces `steebchen/nginx-spa:latest` with the
official Nginx image `nginx:mainline-alpine` using our own
configuration. This configuration allows for paths containing dots to be
redirected to `index.html` like any other path. In addition to that, it
correctly applies gzip compression and Cache-Control headers.