Fix localization strings using nesting
It seems that recently i18next changed its behavior so that strings
using nesting **don't** need to use a top-level namespace when
referencing other strings. This seems to be an undocumented change or
perhaps there is no documentation because it is a regression.

Future i18next package upgrades must be inspected with more caution.